dev/issue-25-validate-colors A few changes: - updated .travis.yml to use the "iu" flags - updated CHANGELOG - updated package.json to have the new version - updated setup.sh to have flags, this will help me test all web components without having to install their packages

+++ setup.sh
@@ -5,13 +5,53 @@ declare -a webComponents=("rms-sparkline-inline" \
 "util-lib"
 )
 
+# Usage
+# - compile components: $ ./setup.sh
+# - install packages and compile components: $ ./setup.sh -i
+# - compile components and run tests: $ ./setup.sh -u
+# - install packages, compile components, and run tests: $ ./setup.sh -iu
+
+# Process script arguments and set control variables
+#
+install="no"
+unitTests="no"
+echo "Process script arguments and set control variables"
+while getopts  "iu" flag
+do
+  echo "$flag" $OPTIND $OPTARG
+
+  if [ $flag = "i" ];
+    then
+      install="yes"
+  fi
+
+
+  if [ $flag = "" ];
+    then
+      unitTests="yes"
+  fi
+
+done
+
 for webComponent in "${webComponents[@]}"
 do
     echo "Initializing " + $webComponent
     pushd $webComponent
-    rm -rf node_modules
-    npm install
+    if [ $install = "yes" ];
+      then
+        echo "Installing packackages"
+        rm -rf node_modules
+        npm install
+      else
+        echo "We WILL NOT install packages"
+    fi
     npm run prepare
-    npm run test
+    if [ $unitTests = "yes" ];
+      then
+        echo "Running unit tests"
+         npm run test
+      else
+        echo "We WILL NOT run unit tests"
+    fi
     popd
 done
